requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.3
